pyproject.toml
[project]
name = "nextjs-fastapi-streaming"
version = "0.1.0"
description = "Next.js FastAPI AI Streaming Application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "openai",
    "python-dotenv",
    "pydantic",
    "mangum>=0.17.0",
]

[tool.uv]
dev-dependencies = []

[project.scripts]
start-api = "uvicorn app.api.index:app --reload --port 8000"
